fix: 修复 event 变更导致的单测失败
