+++ sese/event/BaseEventConvert.h
@@ -2,17 +2,13 @@
 
 #include "sese/event/BaseEvent.h"
 
-namespace sese {
-    namespace event {
-        class BaseEventConvert;
-    }
-}
+namespace sese::event {
+    class BaseEventConvert {
+    public:
+        virtual ~BaseEventConvert() = default;
 
-class sese::event::BaseEventConvert {
-public:
-    virtual ~BaseEventConvert() = default;
+        virtual unsigned int fromNativeEvent(int event) = 0;
 
-    virtual unsigned int fromNativeEvent(int event) = 0;
-
-    virtual int toNativeEvent(unsigned int event) = 0;
-};
+        virtual int toNativeEvent(unsigned int event) = 0;
+    };
+}// namespace sese::event
