--- sese/event/EpollEvent.h
@@ -1,10 +1,0 @@
-#pragma once
-
-#include "sese/event/BaseEvent.h"
-
-namespace sese {
-    namespace event {
-        class EpollEvent : public BaseEvent {
-        };
-    }
-}
