--- sese/event/EpollEventConvert.h
@@ -1,14 +1,0 @@
-#pragma once
-
-#include "sese/event/BaseEventConvert.h"
-
-namespace sese {
-    namespace event {
-        class EpollEventConvert : public BaseEventConvert {
-        public:
-            unsigned int fromNativeEvent(int event) override;
-
-            int toNativeEvent(unsigned int event) override;
-        };
-    }
-}
