+++ sese/event/Event.h
@@ -2,46 +2,40 @@
 
 #if defined(__linux__)
 
-#include "sese/event/EpollEvent.h"
-#include "sese/event/EpollEventLoop.h"
-#include "sese/event/EpollEventConvert.h"
+#include "sese/native/Linux/event/EpollEvent.h"
+#include "sese/native/Linux/event/EpollEventLoop.h"
+#include "sese/native/Linux/event/EpollEventConvert.h"
 
 namespace sese {
-    namespace event {
-        using Event = BaseEvent;
-        using EventLoop = EpollEventLoop;
-        using EventConvert = EpollEventConvert;
-    }
-}
+    using Event = BaseEvent;
+    using EventLoop = EpollEventLoop;
+    using EventConvert = EpollEventConvert;
+}// namespace sese::event
 
 #endif
 
 #if defined(_WIN32)
 
-#include "sese/event/WSAEvent.h"
-#include "sese/event/WSAEventLoop.h"
-#include "sese/event/WSAEventConvert.h"
+#include "sese/native/Windows/event/WSAEvent.h"
+#include "sese/native/Windows/event/WSAEventLoop.h"
+#include "sese/native/Windows/event/WSAEventConvert.h"
 
-namespace sese {
-    namespace event {
-        using Event = BaseEvent;
-        using EventLoop = WSAEventLoop;
-        using EventConvert = WSAEventConvert;
-    }
-}
+namespace sese::event {
+    using Event = BaseEvent;
+    using EventLoop = WSAEventLoop;
+    using EventConvert = WSAEventConvert;
+}// namespace sese::event
 
 #endif
 
 #if defined(__APPLE__)
 
-#include "sese/event/KqueueEvent.h"
-#include "sese/event/KqueueEventLoop.h"
+#include "sese/native/Darwin/event/KqueueEvent.h"
+#include "sese/native/Darwin/event/KqueueEventLoop.h"
 
-namespace sese {
-    namespace event {
-        using Event = BaseEvent;
-        using EventLoop = KqueueEventLoop;
-    }
-}
+namespace sese::event {
+    using Event = BaseEvent;
+    using EventLoop = KqueueEventLoop;
+}// namespace sese::event
 
 #endif
