--- sese/event/KqueueEvent.h
@@ -1,14 +1,0 @@
-#pragma once
-
-#include "sese/event/BaseEvent.h"
-
-namespace sese {
-    namespace event {
-        class KqueueEvent;
-    }
-}
-
-class sese::event::KqueueEvent : public BaseEvent {
-public:
-    unsigned int oldEvents{EVENT_NULL};
-};
