+++ sese/native/Darwin/event/KqueueEvent.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include "sese/event/BaseEvent.h"
+
+namespace sese::event {
+    class KqueueEvent : public BaseEvent {
+    };
+}// namespace sese::event
