+++ sese/native/Linux/event/EpollEvent.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include "sese/event/BaseEvent.h"
+
+namespace sese::event {
+    class EpollEvent : public BaseEvent {
+    };
+}// namespace sese::event
