+++ sese/native/Linux/event/EpollEventConvert.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include "sese/event/BaseEventConvert.h"
+
+namespace sese::event {
+    class EpollEventConvert : public BaseEventConvert {
+    public:
+        unsigned int fromNativeEvent(int event) override;
+
+        int toNativeEvent(unsigned int event) override;
+    };
+}// namespace sese::event
