+++ sese/native/Windows/event/WSAEvent.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include "sese/event/BaseEvent.h"
+
+namespace sese::event {
+    class WSAEvent : public BaseEvent {
+    public:
+        void *wsaEvent{nullptr};
+    };
+}// namespace sese::event
