+++ sese/native/Windows/event/WSAEventConvert.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include "sese/event/BaseEventConvert.h"
+
+namespace sese::event {
+    class WSAEventConvert : public BaseEventConvert {
+    public:
+        unsigned int fromNativeEvent(int event) override;
+
+        int toNativeEvent(unsigned int event) override;
+    };
+}// namespace sese::event
