+++ vcpkg.json
@@ -1,6 +1,6 @@
 {
     "name": "sese-core",
-    "version-string": "1.2.1",
+    "version-string": "1.2.2",
     "supports": "x64",
     "license": "Apache-2.0",
     "dependencies": [
